chore: 프로젝트 코드 정리
- Protocol type을 다른 파일로 빼기
- TabBar -> HomeTabBar로 네이밍 변경
- AuthAPISerType을 채택한 AuthAPIService 타입으로 네이밍 변경
