+++ Boolti/Boolti/Sources/Network/Foundation/NetworkProvider.swift
@@ -1,0 +1,42 @@
+//
+//  NetworkProvider.swift
+//  Boolti
+//
+//  Created by Miro on 1/22/24.
+//
+
+import Foundation
+import RxSwift
+import RxMoya
+import Moya
+
+final class NetworkProvider: Networking {
+
+    private let provider: MoyaProvider<MultiTarget>
+
+    init(plugins: [PluginType] = []) {
+        let session = Session(interceptor: AuthInterceptor())
+        session.sessionConfiguration.timeoutIntervalForRequest = 10
+
+        self.provider = MoyaProvider<MultiTarget>(session: session, plugins: plugins)
+    }
+
+    func request(_ api: BaseAPI) -> Single<Response> {
+        let requestString = "\(api.path)"
+        let endpoint = MultiTarget.target(api)
+
+        return provider.rx.request(endpoint)
+            .filterSuccessfulStatusCodes()
+            .do(
+                onSuccess: { response in
+                    print("SUCCESS: \(requestString) (\(response.statusCode))")
+                },
+                onError: { _ in
+                    print("ERROR: \(requestString)")
+                },
+                onSubscribed: {
+                    print("REQUEST: \(requestString)")
+                }
+            )
+    }
+}
