+++ Boolti/Boolti/Sources/Network/Services/AuthAPIServiceType.swift
@@ -1,0 +1,14 @@
+//
+//  AuthAPIServiceType.swift
+//  Boolti
+//
+//  Created by Miro on 1/22/24.
+//
+
+import Foundation
+import Moya
+import RxSwift
+
+protocol AuthAPIServiceType {
+    func login(accessToken: String) -> Single<KakaoLoginResponseDTO>
+}
