+++ Boolti/Boolti/Sources/UILayer/Splash/SplashDIContainer.swift
@@ -24,7 +24,7 @@ final class SplashDIContainer {
 
     private func createSplashViewModel() -> SplashViewModel {
         // TODO: 네트워크 의존성 주입하는 방법임!! 나중에 지워야함!
-        let viewModel = SplashViewModel(authAPIService: AuthService(provider: self.networkProvider), delegate: self.rootDIContainer.rootViewModel)
+        let viewModel = SplashViewModel(authAPIService: AuthAPIService(provider: self.networkProvider), delegate: self.rootDIContainer.rootViewModel)
 
         return viewModel
     }
